backend/requirements.txt
# backend/requirements.txt - VERSÃO COM GEOJSON (sem fiona/shapely que precisam GDAL)

fastapi>=0.111.0
uvicorn>=0.31.1
gunicorn==22.0.0

pydantic

# GeoJSON (Python puro - não precisa GDAL)
geojson>=3.1.0

# Dependências GEE/STAC + Autenticação Google
earthengine-api>=0.1.419
google-api-python-client
google-auth
google-auth-httplib2
python-dotenv>=1.0.1
requests>=2.28.0

# Google ADK (GRATUITO) + Generative AI
google-adk
google-genai
google-generativeai>=0.8.0